feat: move the configServer features related to caching into a new file
